+++ CHANGELOG.md
@@ -1,6 +1,17 @@
+## [1.0.1+1] - Use flutter stable channel
+
+* Rebuilt example to work with flutter stable
+* Removed shadowColor to make the widget compatible with flutter stable
+* Removed autofillHints to make the widget compatible with flutter stable
+* Removed mouseCursor to make the widget compatible with flutter stable
+* Removed cursorHeight to make the widget compatible with flutter stable
+* Removed obscuringCharacter to make the widget compatible with flutter stable
+* Removed toolbarHeight from AppBar to make the widget compatible with flutter stable
+* Removed leadingWidth to make the widget compatible with flutter stable
+
 ## [1.0.1] - Use flutter stable channel
 
-* Remove shadowColor to make the widget compatible with flutter stable
+* Removed shadowColor to make the widget compatible with flutter stable
 
 ## [1.0.0] - initial release